+++ CHANGELOG.md
@@ -4,6 +4,9 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [unreleased]
 
+- Breaking Change: Remove `DatePicker`.
+- Feature: `StandaloneRangeDatePicker`: Update component and dependencies.
+
 ## [11.6.0] - 2022-07-01
 
 Feature: